+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Cockpit File Sharing 4.3.1-2
+## Cockpit File Sharing 4.3.2-2
 
-* Fix os release display in app info popup on EL8/EL9
-* Enable AD accounts what's new popup for v4.3.1
+* Fix bug causing nfs server crash during export removal with ceph by removing ceph remount after removing share instead of before
+* Fix bug causing NFS hooks to be run multiple times for clustered environments